Refactor Ewald LJ and Electrostatics into a single nonbonded file.
