+++ timemachine/nonbonded_force.py
@@ -1,3 +1,4 @@
+import numpy as np
 import tensorflow as tf
 from timemachine.force import ConservativeForce
 from timemachine.constants import ONE_4PI_EPS0
@@ -12,6 +13,16 @@ def generate_exclusion_masks(exclusions):
     # only use upper triangular part
     exclusions = tf.cast(tf.matrix_band_part(exclusions, 0, -1), dtype=tf.int32)
     return (mask_a - mask_b) - exclusions
+
+
+def generate_inclusion_exclusion_masks(exclusions):
+    ones = tf.ones_like(exclusions, dtype=tf.int32)
+    mask_a = tf.matrix_band_part(ones, 0, -1) # Upper triangular matrix of 0s and 1s
+    mask_b = tf.matrix_band_part(ones, 0, 0) # Diagonal matrix
+
+    # only use upper triangular part
+    exclusions = tf.cast(tf.matrix_band_part(exclusions, 0, -1), dtype=tf.int32)
+    return (mask_a - mask_b) - exclusions, exclusions
 
 
 class LeonnardJonesForce(ConservativeForce):
@@ -87,34 +98,29 @@ class LeonnardJonesForce(ConservativeForce):
 
 class ElectrostaticForce(ConservativeForce):
 
-
-    def __init__(self, params, param_idxs, exclusions):
-        """
-        Implements a non-periodic point charge electrostatic potential.
-
-        Parameters:
-        -----------
-        params: tf.float64 variable
-            charge list, eg. [q_C, q_C+, q_C-, q_H-]
-
-        param_idxs: tf.int32 (N,)    
-            table of indexes into each charge type.
-
-        exclusions: tf.bool (N, N)
-            boolean mask denoting if interaction e[i,j] should be
-            excluded or not. If e[i,j] is 1 then the interaction
-            is excluded, 0 implies it is kept. Note that only the upper
-            right triangular portion of this is used.
-
-        """
-        # (ytz) TODO: implement Ewald/PME
+    def __init__(self, params, param_idxs, exclusions, box=None, kmax=10):
         self.params = params
-        self.param_idxs = param_idxs
-        self.keep_mask = generate_exclusion_masks(exclusions)
-
+        self.param_idxs = param_idxs # length N
+        self.num_atoms = len(self.param_idxs)
+        self.charges = tf.gather(self.params, self.param_idxs)
+        self.charges = tf.reshape(self.charges, shape=(1, -1))
+        self.exclusions = exclusions
+        self.alphaEwald = 1.0
+        self.direct_mask, self.exclusion_mask = generate_inclusion_exclusion_masks(exclusions)
+        self.box = box # we probably want this to be variable when we start to work on barostats
+        self.kmax = kmax
 
     def energy(self, conf):
+        direct_nrg, exclusion_nrg = self.direct_and_exclusion_energy(conf)
+        if self.box is None:
+            return direct_nrg
+        else:
+            return self.reciprocal_energy(conf) + direct_nrg - exclusion_nrg - self.self_energy(conf)
 
+    def self_energy(self, conf):
+        return tf.reduce_sum(ONE_4PI_EPS0 * tf.pow(self.charges, 2) * self.alphaEwald/np.sqrt(np.pi))
+
+    def direct_and_exclusion_energy(self, conf):
         charges = tf.gather(self.params, self.param_idxs)
         qi = tf.expand_dims(charges, 0)
         qj = tf.expand_dims(charges, 1)
@@ -122,16 +128,126 @@ class ElectrostaticForce(ConservativeForce):
 
         ri = tf.expand_dims(conf, 0)
         rj = tf.expand_dims(conf, 1)
-        d2ij = tf.reduce_sum(tf.pow(ri-rj, 2), axis=-1)
 
-        qij_mask = tf.boolean_mask(qij, self.keep_mask)
-        d2ij_mask = tf.boolean_mask(d2ij, self.keep_mask)
+        if self.box is not None:
+            rij = ri - rj
+            base = tf.floor(rij/self.box + 0.5)*self.box # (ytz): can we differentiate through this?
+            dxdydz = tf.pow(rij-base, 2)
+            d2ij = tf.reduce_sum(dxdydz, axis=-1)
+        else:
+            d2ij = tf.reduce_sum(tf.pow(ri-rj, 2), axis=-1)
 
         # (ytz): we move the sqrt to outside of the mask
         # to avoid nans in autograd.
-        eij = qij_mask/tf.sqrt(d2ij_mask)
 
-        return ONE_4PI_EPS0*tf.reduce_sum(eij, axis=-1)
+        # direct
+        qij_direct_mask = tf.boolean_mask(qij, self.direct_mask)
+        d2ij_direct_mask = tf.boolean_mask(d2ij, self.direct_mask)
+        r_direct = tf.sqrt(d2ij_direct_mask)
+        eij_direct = qij_direct_mask/r_direct
+
+        if self.box is not None:
+            # We adjust direct by the erfc, and adjust the reciprocal space's
+            # exclusionary contribution by the direction space weighted by erf
+            eij_direct *= tf.erfc(self.alphaEwald*r_direct)
+            # exclusions to subtract from reciprocal space
+            qij_exclusion_mask = tf.boolean_mask(qij, self.exclusion_mask)
+            d2ij_exclusion_mask = tf.boolean_mask(d2ij, self.exclusion_mask)
+            r_exclusion = tf.sqrt(d2ij_exclusion_mask)
+            eij_exclusion = qij_exclusion_mask/r_exclusion
+            eij_exclusion *= tf.erf(self.alphaEwald*r_exclusion)
+
+            return ONE_4PI_EPS0*tf.reduce_sum(eij_direct, axis=-1), ONE_4PI_EPS0*tf.reduce_sum(eij_exclusion, axis=-1)
+        else:
+            return ONE_4PI_EPS0*tf.reduce_sum(eij_direct, axis=-1), None
+
+    def reciprocal_energy(self, conf):
+        assert self.box is not None
+
+        recipBoxSize = np.array([
+            (2*np.pi)/self.box[0],
+            (2*np.pi)/self.box[1],
+            (2*np.pi)/self.box[2]]
+        )
+
+        mg = []
+        lowry = 0
+        lowrz = 1
+
+        numRx, numRy, numRz = self.kmax, self.kmax, self.kmax
+
+        for rx in range(numRx):
+            for ry in range(lowry, numRy):
+                for rz in range(lowrz, numRz):
+                    mg.append((rx, ry, rz))
+                    lowrz = 1 - numRz
+                lowry = 1 - numRy
+
+        # lattice vectors
+        ki = np.expand_dims(recipBoxSize, axis=0) * mg # [nk, 3]
+        ri = tf.expand_dims(conf, axis=0) # [1, N, 3]
+        rik = tf.reduce_sum(tf.multiply(ri, tf.expand_dims(ki, axis=1)), axis=-1) # [nk, N]
+        real = tf.cos(rik)
+        imag = tf.sin(rik)
+        eikr = tf.complex(real, imag) # [nk, N]
+        qi = tf.complex(self.charges, np.float64(0.0))
+        Sk = tf.reduce_sum(qi*eikr, axis=-1)  # [nk]
+        n2Sk = tf.pow(tf.abs(Sk), 2)
+        k2 = tf.reduce_sum(tf.multiply(ki, ki), axis=-1) # [nk]
+        factorEwald = -1/(4*self.alphaEwald*self.alphaEwald)
+        ak = tf.exp(k2*factorEwald)/k2 # [nk]
+        nrg = tf.reduce_sum(ak * n2Sk)
+        recipCoeff = (ONE_4PI_EPS0*4*np.pi)/(self.box[0]*self.box[1]*self.box[2])
+
+        return recipCoeff * nrg
+
+# class ElectrostaticForce(ConservativeForce):
+
+
+#     def __init__(self, params, param_idxs, exclusions):
+#         """
+#         Implements a non-periodic point charge electrostatic potential.
+
+#         Parameters:
+#         -----------
+#         params: tf.float64 variable
+#             charge list, eg. [q_C, q_C+, q_C-, q_H-]
+
+#         param_idxs: tf.int32 (N,)    
+#             table of indexes into each charge type.
+
+#         exclusions: tf.bool (N, N)
+#             boolean mask denoting if interaction e[i,j] should be
+#             excluded or not. If e[i,j] is 1 then the interaction
+#             is excluded, 0 implies it is kept. Note that only the upper
+#             right triangular portion of this is used.
+
+#         """
+#         # (ytz) TODO: implement Ewald/PME
+#         self.params = params
+#         self.param_idxs = param_idxs
+#         self.keep_mask = generate_exclusion_masks(exclusions)
+
+
+#     def energy(self, conf):
+
+#         charges = tf.gather(self.params, self.param_idxs)
+#         qi = tf.expand_dims(charges, 0)
+#         qj = tf.expand_dims(charges, 1)
+#         qij = tf.multiply(qi, qj)
+
+#         ri = tf.expand_dims(conf, 0)
+#         rj = tf.expand_dims(conf, 1)
+#         d2ij = tf.reduce_sum(tf.pow(ri-rj, 2), axis=-1)
+
+#         qij_mask = tf.boolean_mask(qij, self.keep_mask)
+#         d2ij_mask = tf.boolean_mask(d2ij, self.keep_mask)
+
+#         # (ytz): we move the sqrt to outside of the mask
+#         # to avoid nans in autograd.
+#         eij = qij_mask/tf.sqrt(d2ij_mask)
+
+#         return ONE_4PI_EPS0*tf.reduce_sum(eij, axis=-1)
 
 if __name__ == "__main__":
